Revise workload_generator.py and move helper functions to helpers.py
